\name{wrld}

\alias{wrld}
\alias{merc}


\docType{data}

\title{World countries}

\description{
world coastline and country outlines in longitude/latitude (wrld) and in Mercator projection (merc). 
}

\usage{
data(wrld)
data(merc)
}


\source{
Derived from the wrld_simpl data set in package maptools
}

\keyword{datasets}

