\name{finalBrng}
\Rdversion{1.1}
\alias{finalBrng}

\title{
final bearing
}

\description{
Get the final bearing when arriving at a point
}

\usage{
finalBrng(p, brng, d, R = 6378137)
}

\arguments{
  \item{p}{longitude/latitude of point(s) 1; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object}
  \item{brng}{bearing}
  \item{d}{distance}
  \item{R}{Radius of the earth}
}


\details{
}

\value{

}

\references{
\url{http://www.movable-type.co.uk/scripts/latlong-vincenty.html}
}

\author{
Chris Veness
}
\note{
}


\seealso{
}
\examples{
}

\keyword{ spatial  }

