\name{bearing}
\Rdversion{1.1}

\alias{bearing}

\title{
Bearing
}

\description{
Get the bearing between two points
}

\usage{
bearing(p1, p2)
}

\arguments{
  \item{p1}{longitude/latitude of point(s) 1; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object}
  \item{p2}{longitude/latitude of point(s) 2; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object}
}

\details{
}

\value{
}


\references{
}

\author{
Chris Veness
}



\seealso{
}

\examples{
}

\keyword{ spatial }

