\name{brngRhumb}
\Rdversion{1.1}

\alias{brngRhumb}

\title{
brngRhumb
}


\description{
Bearing along the rhumb line
}

\usage{
brngRhumb(p1, p2)
}

\arguments{
  \item{p1}{longitude/latitude of point(s) 1; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object}
  \item{p2}{as above. Should be of same length of p1, or a single point (or vice versa when p1 is a single point}
}

\details{
}

\value{
...
}

\references{
\url{http://www.movable-type.co.uk/scripts/latlong.html}
}

\author{
Chris Veness
}

\note{
}


\seealso{
}

\examples{
}

\keyword{ spatial }

