\name{geosphere-package}
\alias{geosphere-package}
\alias{geosphere}
\docType{package}
\title{Spherical Trigonometry}

\description{
Spherical trigonometry for geographic applications, such as great circle distance and distance along a rhumb line.
}

\details{
\tabular{ll}{
Package: \tab geosphere\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 6-Feb-2010\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}

The functions compute distance and bearing (course) along great circles (= shortest distance on a sphere, or "as the crow flies") and rhumb lines (lines of constant bearing). Geographic locations are always specified in degrees (NOT radians), and the default unit of distance is meter.

This is the first version of this package, so please look out for bugs and let us know if you find any.
}

\author{
Robert Hijmans, Ed Williams, Chris Veness

Maintainer: Robert J. Hijmans <r.hijmans@gmail.com>
}

\references{
\url{http://williams.best.vwh.net/avform.htm}

\url{http://www.movable-type.co.uk/scripts/latlong.html}

\url{http://en.wikipedia.org/wiki/Great_circle_distance}

\url{http://mathworld.wolfram.com/SphericalTrigonometry.html}
}

\keyword{ package }
\keyword{ spatial }

