\name{geosphere-package}
\alias{geosphere-package}
\alias{geosphere}
\docType{package}
\title{Spherical Trigonometry}

\description{
Spherical trigonometry for geographic applications.

The functions in this package compute distance and bearing (course) along great circles (= shortest distance on a sphere, or "as the crow flies") and rhumb lines (lines of constant bearing). Geographic locations are always specified in degrees (NOT radians), and the default unit of distance is meter.

This is the first version of this package, so please look out for bugs and let us know if you find any.
}

\author{
Robert Hijmans, Ed Williams, Chris Veness

Maintainer: Robert J. Hijmans <r.hijmans@gmail.com>
}

\references{
\url{http://williams.best.vwh.net/avform.htm}

\url{http://www.movable-type.co.uk/scripts/latlong.html}

\url{http://en.wikipedia.org/wiki/Great_circle_distance}

\url{http://mathworld.wolfram.com/SphericalTrigonometry.html}
}

\section{Acknowledgements}{
We thank the David Purdy for suggestions to improve the package.
}


\keyword{ package }
\keyword{ spatial }

