\name{geosphere-package}
\alias{geosphere-package}
\alias{geosphere}
\docType{package}
\title{Spherical Trigonometry}

\description{
Spherical trigonometry for geographic applications, such as great circle distance and distance along a rhumb line.
}

\details{
\tabular{ll}{
Package: \tab geosphere\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2009-10-10\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Robert Hijmans, Ed Williams, Chris Veness

Maintainer: Robert J. Hijmans <r.hijmans@gmail.com>
}

\references{
\url{http://williams.best.vwh.net/avform.htm}

\url{http://www.movable-type.co.uk/scripts/latlong.html}

\url{http://en.wikipedia.org/wiki/Great_circle_distance}

\url{http://mathworld.wolfram.com/SphericalTrigonometry.html}
}


\keyword{ package }
\keyword{ spatial }
