\name{geosphere-package}
\alias{geosphere-package}
\alias{geosphere}
\docType{package}
\title{
What the package does (short line)
Spherical Trigonometry
}

\description{
Spherical Trigonometry for geographic applications. Including Great Circle distance, ...
}

\details{
\tabular{ll}{
Package: \tab geosphere\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-10\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}

~~ An overview of how to use the package, including the most important functions ~~

}
\author{
Chris Veness

Maintainer: Robert J. Hijmans <r.hijmans@gmail.com>
}
\references{

}
\keyword{ package }
\keyword{ spatial }

\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
}
